Compatibility for 2022
Back in April 2022, Rockstar removed the IS_BIT_SET native and instead replaced it with an opcode within the VM. This update adds compatibility for said change.

In addition, I've removed the "magic" check as it causes more difficulties than it's worth. Going forward I am considering making this compatible with Red Dead Redemption 2. That in itself is a good reason for not paying attention to the "magic" key